refactor: split config entry setup phases
