+++ custom_components/helianthus/entry_setup.py
@@ -1,0 +1,1750 @@
+"""Config-entry setup orchestration for the Helianthus integration."""
+
+from __future__ import annotations
+
+import re
+from urllib.parse import urlsplit
+
+from . import (
+    CONF_DHW_SCHEDULE_HELPER,
+    CONF_INSTANCE_GUID,
+    CONF_PATH,
+    CONF_TRANSPORT,
+    CONF_USE_SUBSCRIPTIONS,
+    CONF_VERSION,
+    CONF_ZONE_SCHEDULE_HELPERS,
+    DEFAULT_DHW_SCHEDULE_HELPER,
+    DEFAULT_GRAPHQL_PATH,
+    DEFAULT_GRAPHQL_TRANSPORT,
+    DEFAULT_SCAN_INTERVAL,
+    DEFAULT_USE_SUBSCRIPTIONS,
+    DEFAULT_ZONE_SCHEDULE_HELPERS,
+    DOMAIN,
+    PLATFORMS,
+    _LOGGER,
+    _canonical_bus_display_name,
+    _canonical_bus_model_name,
+    _clean_label,
+    _config_entry_sort_key,
+    _entry_identity_probe_addresses,
+    _find_verified_entry_by_configured_instance_guid,
+    _format_hex4_version,
+    _iter_identifier_pairs,
+    _migrate_unique_ids_to_snake_case,
+    _parse_zone_schedule_helper_bindings,
+    _select_bus_migration_target,
+    _update_entry_endpoint_if_changed,
+    async_sanitize_legacy_eebus_admin_entry,
+    async_setup_eebus_admin_boundary,
+    async_unload_eebus_admin_boundary,
+    update_effective_admission,
+)
+
+
+async def _async_forward_platforms_and_finalize(
+    hass: object,
+    entry: object,
+    platforms: list[str],
+    *,
+    cleanup_nonexportable_radio_registry_entries: object,
+    nonexportable_radio_keys: set[str],
+    run_trusted_cleanup: object,
+) -> None:
+    """Forward platforms before the two established post-setup cleanup passes."""
+    await hass.config_entries.async_forward_entry_setups(entry, platforms)
+    cleanup_nonexportable_radio_registry_entries(
+        nonexportable_radio_keys,
+        "post-platform setup",
+    )
+    run_trusted_cleanup("post-platform setup")
+
+
+async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
+    """Set up Helianthus from a config entry."""
+    from homeassistant.core import callback
+    from homeassistant.helpers import device_registry as dr
+    from homeassistant.helpers import entity_registry as er
+    from homeassistant.helpers.aiohttp_client import async_get_clientsession
+    from homeassistant.helpers.event import async_track_state_change_event
+    from homeassistant.const import CONF_HOST, CONF_PORT, CONF_SCAN_INTERVAL
+    from .graphql import GraphQLClient, build_graphql_url
+    from .identity import (
+        GatewayIdentityVerificationError,
+        configured_instance_guid,
+        normalize_instance_guid,
+        updated_entry_data,
+        verify_gateway_identity,
+    )
+    from .coordinator import (
+        HelianthusAdapterInfoCoordinator,
+        HelianthusBoilerCoordinator,
+        HelianthusCircuitCoordinator,
+        HelianthusCoordinator,
+        HelianthusEnergyCoordinator,
+        HelianthusFM5Coordinator,
+        HelianthusRadioDeviceCoordinator,
+        HelianthusScheduleCoordinator,
+        HelianthusSemanticCoordinator,
+        HelianthusSystemCoordinator,
+        HelianthusStatusCoordinator,
+    )
+    from .device_ids import (
+        adapter_identifier,
+        boiler_burner_identifier,
+        boiler_hydraulics_identifier,
+        build_radio_bus_key,
+        build_bus_device_key,
+        bus_identifier,
+        circuit_display_name,
+        circuit_identifier,
+        circuit_type_display_name,
+        daemon_identifier,
+        exportable_radio_bus_keys,
+        has_bus_identity_evidence,
+        is_b524_inventory_radio_bus_key,
+        managing_device_identifier,
+        nonexportable_radio_bus_keys,
+        radio_device_identifier,
+        radio_bus_key_from_device,
+        resolve_bus_address,
+        resolve_boiler_physical_device_id,
+        resolve_boiler_via_device_id,
+        should_export_radio_device,
+        should_remove_missing_radio_bus_key,
+        stable_bus_identity_model,
+    )
+    from .subscriptions import start_subscriptions
+    from .pv_m2m import (
+        async_setup_pv_m2m_boundary,
+        pv_m2m_option_signature,
+    )
+    from .entry_services import async_setup_optional_eebus_admin_service
+    from .zone_parent import (
+        build_zone_parent_device_ids,
+        radio_mappings_by_zone_id,
+        resolve_retained_parent_bindings,
+        should_reload_zone_parent_state,
+        zone_ids_by_climate_unique_id,
+    )
+
+    await async_sanitize_legacy_eebus_admin_entry(hass, entry)
+
+    device_registry = dr.async_get(hass)
+    entity_registry = er.async_get(hass)
+
+    def refuse_duplicate_setup(owner_entry: object, instance_guid: str) -> bool:
+        options_merged = _merge_duplicate_config_entry_options(
+            hass,
+            alias_entry=entry,
+            owner_entry=owner_entry,
+        )
+        removed_entities, removed_devices = _remove_config_entry_registry_state(
+            device_registry=device_registry,
+            entity_registry=entity_registry,
+            entry_id=entry.entry_id,
+            device_entries_for_config_entry=dr.async_entries_for_config_entry,
+            entity_entries_for_config_entry=er.async_entries_for_config_entry,
+        )
+        _LOGGER.warning(
+            "Refusing to set up Helianthus entry %s because %s reports "
+            "instance GUID %s already owned by entry %s; removed %d stale "
+            "entities and %d stale devices; migrated_options=%s",
+            entry.entry_id,
+            host,
+            instance_guid,
+            owner_entry.entry_id,
+            removed_entities,
+            removed_devices,
+            options_merged,
+        )
+        if _schedule_duplicate_config_entry_removal(hass, entry.entry_id):
+            _LOGGER.warning(
+                "Scheduled duplicate Helianthus config entry %s for removal",
+                entry.entry_id,
+            )
+        return False
+
+    # Phase C2: migrate camelCase unique_ids to snake_case before platform setup.
+    uid_migrated = _migrate_unique_ids_to_snake_case(hass, entry)
+    if uid_migrated:
+        _LOGGER.info(
+            "Migrated %d entity unique_ids from camelCase to snake_case for entry %s",
+            uid_migrated,
+            entry.entry_id,
+        )
+
+    active_entry_ids = {
+        config_entry.entry_id for config_entry in hass.config_entries.async_entries(DOMAIN)
+    }
+
+    stale_entities_removed = 0
+    for entity_entry in tuple(entity_registry.entities.values()):
+        if entity_entry.platform != DOMAIN:
+            continue
+        if entity_entry.config_entry_id in active_entry_ids:
+            continue
+        entity_registry.async_remove(entity_entry.entity_id)
+        stale_entities_removed += 1
+
+    stale_devices_removed = 0
+    for device_entry in tuple(device_registry.devices.values()):
+        identifier_pairs = _iter_identifier_pairs(device_entry.identifiers)
+        domain_tokens = [
+            token
+            for identifier_domain, token in identifier_pairs
+            if identifier_domain == DOMAIN
+        ]
+        if not domain_tokens:
+            continue
+        if any(_identifier_matches_any_entry(token, active_entry_ids) for token in domain_tokens):
+            continue
+        device_registry.async_remove_device(device_entry.id)
+        stale_devices_removed += 1
+
+    if stale_entities_removed or stale_devices_removed:
+        _LOGGER.info(
+            "Helianthus stale cleanup removed %d entities and %d devices",
+            stale_entities_removed,
+            stale_devices_removed,
+        )
+
+    host = entry.data.get(CONF_HOST)
+    port = entry.data.get(CONF_PORT)
+    if not host or not port:
+        _LOGGER.warning("Config entry missing host/port; device discovery skipped")
+        return True
+
+    session = async_get_clientsession(hass)
+    path = entry.data.get(CONF_PATH) or DEFAULT_GRAPHQL_PATH
+    transport = entry.data.get(CONF_TRANSPORT) or DEFAULT_GRAPHQL_TRANSPORT
+    version = (entry.data.get(CONF_VERSION) or "").strip() or None
+    entry_instance_guid = configured_instance_guid(entry.data, entry.unique_id)
+    try:
+        verified_endpoint = await verify_gateway_identity(
+            session=session,
+            host=str(host),
+            port=int(port),
+            path=path,
+            transport=transport,
+            addresses=_entry_identity_probe_addresses(entry.data, str(host)),
+            expected_instance_guid=None,
+            version=version,
+        )
+    except (GatewayIdentityVerificationError, TypeError, ValueError) as exc:
+        reason = getattr(exc, "reason", type(exc).__name__)
+        _LOGGER.debug(
+            "Stable identity refresh skipped for Helianthus entry %s: %s",
+            entry.entry_id,
+            reason,
+        )
+    else:
+        live_instance_guid = verified_endpoint.instance_guid
+        if entry_instance_guid is None:
+            entry_instance_guid = live_instance_guid
+            hass.config_entries.async_update_entry(
+                entry,
+                data=updated_entry_data(
+                    entry.data,
+                    verified_endpoint,
+                    version=version or verified_endpoint.version,
+                ),
+                unique_id=entry_instance_guid,
+            )
+            host = verified_endpoint.host
+            port = verified_endpoint.port
+            path = verified_endpoint.path
+            transport = verified_endpoint.transport
+            version = version or verified_endpoint.version
+            _LOGGER.info(
+                "Migrated Helianthus entry %s to stable instance GUID %s",
+                entry.entry_id,
+                entry_instance_guid,
+            )
+        elif live_instance_guid != entry_instance_guid:
+            duplicate_entry = await _find_verified_entry_by_configured_instance_guid(
+                hass,
+                session,
+                live_instance_guid,
+                exclude_entry_id=entry.entry_id,
+            )
+            if duplicate_entry is not None:
+                return refuse_duplicate_setup(duplicate_entry, live_instance_guid)
+
+            entry_instance_guid = live_instance_guid
+            hass.config_entries.async_update_entry(
+                entry,
+                data=updated_entry_data(
+                    entry.data,
+                    verified_endpoint,
+                    version=version or verified_endpoint.version,
+                ),
+                unique_id=entry_instance_guid,
+            )
+            host = verified_endpoint.host
+            port = verified_endpoint.port
+            path = verified_endpoint.path
+            transport = verified_endpoint.transport
+            version = version or verified_endpoint.version
+            _LOGGER.warning(
+                "Rebound Helianthus entry %s from stale instance GUID to live GUID %s",
+                entry.entry_id,
+                entry_instance_guid,
+            )
+        else:
+            duplicate_entry = await _find_verified_entry_by_configured_instance_guid(
+                hass,
+                session,
+                live_instance_guid,
+                exclude_entry_id=entry.entry_id,
+            )
+            if (
+                duplicate_entry is not None
+                and _config_entry_sort_key(duplicate_entry)
+                < _config_entry_sort_key(entry)
+            ):
+                return refuse_duplicate_setup(duplicate_entry, live_instance_guid)
+            if _update_entry_endpoint_if_changed(
+                hass,
+                entry,
+                verified_endpoint,
+                version=version or verified_endpoint.version,
+            ):
+                host = verified_endpoint.host
+                port = verified_endpoint.port
+                path = verified_endpoint.path
+                transport = verified_endpoint.transport
+                version = version or verified_endpoint.version
+                _LOGGER.info(
+                    "Rebound Helianthus entry %s to reachable endpoint %s:%s",
+                    entry.entry_id,
+                    host,
+                    port,
+                )
+
+    if entry_instance_guid is not None:
+        normalized_unique_id = normalize_instance_guid(entry.unique_id)
+        stored_data_guid = normalize_instance_guid(entry.data.get(CONF_INSTANCE_GUID))
+        if normalized_unique_id != entry_instance_guid or stored_data_guid != entry_instance_guid:
+            updated_data = dict(entry.data)
+            updated_data[CONF_INSTANCE_GUID] = entry_instance_guid
+            hass.config_entries.async_update_entry(
+                entry,
+                data=updated_data,
+                unique_id=entry_instance_guid,
+            )
+
+    daemon_device_id = daemon_identifier(entry.entry_id)
+    adapter_device_id = adapter_identifier(entry.entry_id)
+    existing_entry_devices = tuple(dr.async_entries_for_config_entry(device_registry, entry.entry_id))
+
+    daemon_device_entry = device_registry.async_get_or_create(
+        config_entry_id=entry.entry_id,
+        identifiers={daemon_device_id},
+        manufacturer="Helianthus",
+        model="Helianthus Daemon",
+        name="Helianthus Daemon",
+    )
+
+    adapter_device_entry = device_registry.async_get_or_create(
+        config_entry_id=entry.entry_id,
+        identifiers={adapter_device_id},
+        manufacturer="Helianthus",
+        model="eBUS Adapter",
+        name="eBUS Adapter",
+        via_device=daemon_device_id,
+    )
+    adapter_registry_device_id = adapter_device_entry.id
+    del daemon_device_entry
+
+    graphql_url = build_graphql_url(host, port, path=path, transport=transport)
+    client = GraphQLClient(session=session, url=graphql_url)
+    scan_interval = entry.options.get(CONF_SCAN_INTERVAL, DEFAULT_SCAN_INTERVAL)
+    device_coordinator = HelianthusCoordinator(hass, client, scan_interval)
+    status_coordinator = HelianthusStatusCoordinator(hass, client, scan_interval)
+    semantic_coordinator = HelianthusSemanticCoordinator(hass, client, scan_interval)
+    energy_coordinator = HelianthusEnergyCoordinator(hass, client, scan_interval)
+    circuit_coordinator = HelianthusCircuitCoordinator(hass, client, scan_interval)
+    radio_coordinator = HelianthusRadioDeviceCoordinator(hass, client, scan_interval)
+    fm5_coordinator = HelianthusFM5Coordinator(hass, client, scan_interval)
+    system_coordinator = HelianthusSystemCoordinator(hass, client, scan_interval)
+    boiler_coordinator = HelianthusBoilerCoordinator(hass, client, scan_interval)
+    schedule_coordinator = HelianthusScheduleCoordinator(hass, client, scan_interval)
+    adapter_info_coordinator = HelianthusAdapterInfoCoordinator(hass, client, scan_interval)
+    await device_coordinator.async_config_entry_first_refresh()
+    await status_coordinator.async_config_entry_first_refresh()
+    await semantic_coordinator.async_config_entry_first_refresh()
+    await energy_coordinator.async_config_entry_first_refresh()
+    await circuit_coordinator.async_config_entry_first_refresh()
+    await radio_coordinator.async_config_entry_first_refresh()
+    await fm5_coordinator.async_config_entry_first_refresh()
+    await system_coordinator.async_config_entry_first_refresh()
+    await boiler_coordinator.async_config_entry_first_refresh()
+    await schedule_coordinator.async_config_entry_first_refresh()
+    await adapter_info_coordinator.async_config_entry_first_refresh()
+
+    # AdminV1 is optional and isolated from the GraphQL session. Its failure
+    # must leave the primary integration available with diagnostics only.
+    parsed_graphql_url = urlsplit(graphql_url)
+    admin_origin = f"{parsed_graphql_url.scheme}://{parsed_graphql_url.netloc}"
+    admin_coordinator, admin_session, admin_services_registered = (
+        await async_setup_optional_eebus_admin_service(
+            hass,
+            entry_id=entry.entry_id,
+            origin=admin_origin,
+            instance_guid=entry_instance_guid or entry.entry_id,
+            scan_interval=scan_interval,
+        )
+    )
+
+    adapter_hw = adapter_info_coordinator.data
+    if isinstance(adapter_hw, dict) and adapter_hw.get("firmware_version"):
+        hw_id = adapter_hw.get("hardware_id") or None
+        device_registry.async_update_device(
+            adapter_device_entry.id,
+            sw_version=adapter_hw["firmware_version"],
+            hw_version=hw_id,
+            serial_number=hw_id,
+        )
+
+    devices = device_coordinator.data or []
+    reload_scheduled = False
+
+    def schedule_reload(reason: str) -> None:
+        nonlocal reload_scheduled
+        if reload_scheduled:
+            return
+        reload_scheduled = True
+        _LOGGER.info("Reloading Helianthus entry %s: %s", entry.entry_id, reason)
+        hass.async_create_task(hass.config_entries.async_reload(entry.entry_id))
+
+    def resolved_bus_device_key(device: dict) -> str | None:
+        if not has_bus_identity_evidence(device):
+            return None
+        address = resolve_bus_address(device.get("address"), device.get("addresses"))
+        if address is None:
+            return None
+        model = stable_bus_identity_model(device.get("device_id"), device.get("product_model"))
+        return build_bus_device_key(
+            model=model,
+            address=address,
+            serial_number=_clean_label(device.get("serial_number")),
+            mac_address=_clean_label(device.get("mac_address")),
+            hardware_version=_clean_label(device.get("hardware_version")),
+            software_version=_clean_label(device.get("software_version")),
+        )
+
+    def is_regulator_device(device: dict) -> bool:
+        device_id = str(device.get("device_id") or "").upper()
+        display_name = str(device.get("display_name") or "").upper()
+        product_family = str(device.get("product_family") or "").upper()
+        return bool(
+            device_id.startswith("BASV")
+            or device_id.startswith("VRC")
+            or "SENSOCOMFORT" in display_name
+            or "SENSOCOMFORT" in product_family
+        )
+
+    def parse_circuit_index(circuit: dict) -> int | None:
+        index = circuit.get("index")
+        if isinstance(index, bool):
+            return None
+        try:
+            parsed = int(index)
+        except (TypeError, ValueError):
+            return None
+        if parsed < 0:
+            return None
+        return parsed
+
+    def parse_optional_int(value: object | None) -> int | None:
+        if isinstance(value, bool):
+            return None
+        try:
+            return int(value)
+        except (TypeError, ValueError):
+            return None
+
+    def radio_model_name(device: dict) -> str:
+        model = _clean_label(device.get("device_model"))
+        if model:
+            return model
+        class_address = parse_optional_int(device.get("device_class_address"))
+        if class_address == 0x15:
+            return "VRC720f/2"
+        if class_address == 0x35:
+            return "VR92f"
+        if class_address == 0x26:
+            return "VR71/FM5"
+        if class_address is not None and class_address >= 0:
+            return f"Unknown Radio (0x{class_address:02X})"
+        return "Unknown Radio"
+
+    def radio_firmware_display(value: object | None) -> str | None:
+        token = _clean_label(value)
+        if not token:
+            return None
+        compact = token.replace(".", "")
+        if len(compact) == 6 and compact.isdigit():
+            major = compact[0:2]
+            minor = compact[2:4]
+            patch = compact[4:6]
+            if patch == "00":
+                return f"{major}.{minor}"
+            return f"{major}.{minor}.{patch}"
+        if len(compact) == 4 and compact.isdigit():
+            return f"{compact[0:2]}.{compact[2:4]}"
+        if "." in token:
+            parts = [part.strip() for part in token.split(".") if part.strip()]
+            if len(parts) >= 2:
+                major = parts[0]
+                minor = parts[1]
+                patch = parts[2] if len(parts) >= 3 else ""
+                if patch in {"", "0", "00"}:
+                    return f"{major}.{minor}"
+                return f"{major}.{minor}.{patch}"
+        return token
+
+    known_bus_devices: set[str] = set()
+    bus_address_device_ids: dict[int, tuple[str, str]] = {}
+    regulator_device: tuple[str, str] | None = None
+    regulator_bus_address: int | None = None
+    boiler_device: tuple[str, str] | None = None
+    vr71_device: tuple[str, str] | None = None
+    for device in devices:
+        address = resolve_bus_address(device.get("address"), device.get("addresses"))
+        device_id = device.get("device_id", "unknown")
+        serial_number = device.get("serial_number")
+        manufacturer = _clean_label(device.get("manufacturer")) or "Unknown"
+        sw_version = device.get("software_version")
+        hw_version = device.get("hardware_version")
+        display_name = _canonical_bus_display_name(device)
+
+        if address is None:
+            continue
+
+        bus_device_key = resolved_bus_device_key(device)
+        if not bus_device_key:
+            continue
+        known_bus_devices.add(bus_device_key)
+
+        device_name = display_name or f"{manufacturer} {device_id}"
+        model_name = _canonical_bus_model_name(device)
+
+        bus_device_id = bus_identifier(entry.entry_id, bus_device_key)
+        migration_target = _select_bus_migration_target(
+            existing_entry_devices,
+            entry_id=entry.entry_id,
+            stable_identifier=bus_device_id,
+            address=address,
+            manufacturer=manufacturer,
+            model_name=model_name,
+            serial_number=_clean_label(serial_number),
+        )
+        if migration_target is not None:
+            update_kwargs: dict[str, object] = {
+                "merge_identifiers": {bus_device_id},
+                "manufacturer": manufacturer,
+                "model": model_name,
+                "name": device_name,
+                "via_device_id": adapter_registry_device_id,
+            }
+            if serial_number:
+                update_kwargs["serial_number"] = str(serial_number)
+            if sw_version:
+                update_kwargs["sw_version"] = _format_hex4_version(str(sw_version))
+            if hw_version:
+                update_kwargs["hw_version"] = hw_version
+            device_registry.async_update_device(migration_target.id, **update_kwargs)
+        else:
+            device_kwargs = {
+                "config_entry_id": entry.entry_id,
+                "identifiers": {bus_device_id},
+                "manufacturer": manufacturer,
+                "model": model_name,
+                "name": device_name,
+                "via_device": adapter_device_id,
+            }
+            if serial_number:
+                device_kwargs["serial_number"] = str(serial_number)
+            if sw_version:
+                device_kwargs["sw_version"] = _format_hex4_version(str(sw_version))
+            if hw_version:
+                device_kwargs["hw_version"] = hw_version
+            device_registry.async_get_or_create(**device_kwargs)
+        bus_address_device_ids[address] = bus_device_id
+
+        device_id_upper = str(device_id).upper()
+        if boiler_device is None and device_id_upper.startswith("BAI"):
+            boiler_device = bus_device_id
+        if vr71_device is None and (
+            device_id_upper.startswith("VR_71") or device_id_upper.startswith("VR71")
+        ):
+            vr71_device = bus_device_id
+        if device_id_upper.startswith("BASV"):
+            regulator_device = bus_device_id
+            regulator_bus_address = address
+        elif regulator_device is None and is_regulator_device(device):
+            regulator_device = bus_device_id
+            regulator_bus_address = address
+
+    # Extract regulator manufacturer for virtual device entities.
+    regulator_manufacturer = "Helianthus"
+    for device in devices:
+        if is_regulator_device(device):
+            mfr = _clean_label(device.get("manufacturer"))
+            if mfr:
+                regulator_manufacturer = mfr
+                break
+
+    radio_payload = radio_coordinator.data or {}
+    raw_radio_devices_payload = radio_payload.get("radio_devices")
+    if not isinstance(raw_radio_devices_payload, list):
+        raw_radio_devices_payload = []
+    nonexportable_radio_keys = nonexportable_radio_bus_keys(raw_radio_devices_payload)
+    radio_devices_payload = [
+        device
+        for device in raw_radio_devices_payload
+        if isinstance(device, dict) and should_export_radio_device(device)
+    ]
+
+    # ADR-027: Merge B524 function-module enrichment into physical bus devices.
+    # Group 0x0C entries whose deviceClassAddress matches a known bus address
+    # are enrichment metadata for that bus device, not standalone devices.
+    _B524_FUNCTION_MODULE_GROUP = 0x0C
+    b524_merge_targets: dict[str, tuple[str, str]] = {}
+    for device in radio_devices_payload:
+        group = parse_optional_int(device.get("group"))
+        if group != _B524_FUNCTION_MODULE_GROUP:
+            continue
+        class_addr = parse_optional_int(device.get("device_class_address"))
+        if class_addr is None:
+            continue
+        bus_device_id = bus_address_device_ids.get(class_addr)
+        if bus_device_id is None:
+            continue
+        bus_key = radio_bus_key_from_device(device)
+        if bus_key is None:
+            continue
+        b524_merge_targets[bus_key] = bus_device_id
+        _LOGGER.debug(
+            "B524 merge: radio slot %s (group=0x%02X, class_addr=%d) -> bus device %s",
+            bus_key, group, class_addr, bus_device_id,
+        )
+
+    known_radio_bus_keys: set[str] = set()
+    radio_parent = regulator_device or adapter_device_id
+    for device in radio_devices_payload:
+        group = parse_optional_int(device.get("group"))
+        instance = parse_optional_int(device.get("instance"))
+        if group is None or instance is None:
+            continue
+        bus_key = radio_bus_key_from_device(device)
+        if bus_key is None:
+            continue
+        known_radio_bus_keys.add(bus_key)
+        # ADR-027: skip HA device creation for merged B524 function module slots.
+        if bus_key in b524_merge_targets:
+            continue
+        radio_device_id = radio_device_identifier(entry.entry_id, bus_key)
+        model_name = radio_model_name(device)
+        device_kwargs: dict[str, object] = {
+            "config_entry_id": entry.entry_id,
+            "identifiers": {radio_device_id},
+            "manufacturer": "Vaillant",
+            "model": model_name,
+            "name": model_name,
+        }
+        if radio_parent is not None:
+            device_kwargs["via_device"] = radio_parent
+        sw_version = radio_firmware_display(device.get("firmware_version"))
+        if sw_version:
+            device_kwargs["sw_version"] = sw_version
+        hardware_identifier = parse_optional_int(device.get("hardware_identifier"))
+        if hardware_identifier is not None and hardware_identifier >= 0:
+            device_kwargs["hw_version"] = f"0x{hardware_identifier:04X}"
+        device_registry.async_get_or_create(**device_kwargs)
+
+    circuits_payload = circuit_coordinator.data or {}
+    circuits = circuits_payload.get("circuits", []) or []
+    known_circuit_indexes: set[int] = set()
+    for circuit in circuits:
+        if not isinstance(circuit, dict):
+            continue
+        index = parse_circuit_index(circuit)
+        if index is None:
+            continue
+        known_circuit_indexes.add(index)
+        circuit_type_name = circuit_type_display_name(circuit.get("circuit_type"))
+        managing_device = managing_device_identifier(
+            group=0x02,
+            instance=index,
+            regulator_device_id=regulator_device,
+            vr71_device_id=vr71_device,
+            adapter_device_id=adapter_device_id,
+            managing_device=circuit.get("managing_device"),
+        )
+        circuit_device_id = circuit_identifier(entry.entry_id, index)
+        device_kwargs: dict[str, object] = {
+            "config_entry_id": entry.entry_id,
+            "identifiers": {circuit_device_id},
+            "manufacturer": regulator_manufacturer,
+            "model": circuit_type_name,
+            "name": circuit_display_name(circuit, index),
+        }
+        if managing_device is not None:
+            device_kwargs["via_device"] = managing_device
+        device_registry.async_get_or_create(**device_kwargs)
+
+    fm5_payload = fm5_coordinator.data or {}
+    known_fm5_mode = str(fm5_payload.get("fm5_semantic_mode") or "ABSENT").strip().upper()
+    known_cylinder_indexes: set[int] = set()
+    for cylinder in fm5_payload.get("cylinders", []) or []:
+        if not isinstance(cylinder, dict):
+            continue
+        index = parse_optional_int(cylinder.get("index"))
+        if index is not None and index >= 0:
+            known_cylinder_indexes.add(index)
+
+    admission = update_effective_admission(status_coordinator, devices)
+
+    cleanup_counts = hass.data.setdefault(DOMAIN, {}).setdefault("_admission_cleanup_counts", {})
+    cleanup_allowed = False
+    cleanup_ran = False
+
+    def record_cleanup_evidence(current_devices: list[dict]) -> bool:
+        nonlocal cleanup_allowed
+        update_effective_admission(status_coordinator, current_devices)
+        if status_admission_trusted(status_coordinator) and current_devices:
+            cleanup_counts[entry.entry_id] = int(cleanup_counts.get(entry.entry_id) or 0) + 1
+        else:
+            cleanup_counts[entry.entry_id] = 0
+        cleanup_allowed = cleanup_counts.get(entry.entry_id, 0) >= 2
+        return cleanup_allowed
+
+    record_cleanup_evidence(devices)
+
+    semantic = semantic_coordinator.data or {}
+    semantic_zones = [
+        zone for zone in (semantic.get("zones", []) or []) if isinstance(zone, dict)
+    ]
+    known_zones: set[str] = {
+        str(zone.get("id"))
+        for zone in semantic_zones
+        if zone.get("id") is not None
+    }
+    known_has_dhw = semantic.get("dhw") is not None
+
+    radio_sensor_unique_id_re = re.compile(
+        rf"^{re.escape(entry.entry_id)}-radio-(?P<group>[0-9a-f]{{2}})-(?P<instance>\d{{2}})-sensor-(?P<key>.+)$"
+    )
+    radio_connected_unique_id_re = re.compile(
+        rf"^{re.escape(entry.entry_id)}-radio-(?P<group>[0-9a-f]{{2}})-(?P<instance>\d{{2}})-connected$"
+    )
+    solar_unique_id_re = re.compile(
+        rf"^{re.escape(entry.entry_id)}-solar-(?P<kind>sensor|binary)-(?P<key>.+)$"
+    )
+    cylinder_unique_id_re = re.compile(
+        rf"^{re.escape(entry.entry_id)}-cylinder-(?P<index>\d+)-(?:(?:config-(?P<config_key>.+))|(?P<kind>temperature))$"
+    )
+    status_sparse_unique_ids = frozenset(
+        {
+            f"daemon-{entry.entry_id}-admission_repair_code",
+        }
+    )
+    boiler_temp_unique_id_re = re.compile(
+        rf"^{re.escape(entry.entry_id)}-boiler-(?P<key>(?:flow|return|dhw|dhw_storage)_temperature_c)$"
+    )
+    boiler_sensor_sparse_unique_id_re = re.compile(
+        rf"^{re.escape(entry.entry_id)}-boiler-sensor-(?P<key>hours_till_service)$"
+    )
+    boiler_number_unique_id_re = re.compile(
+        rf"^{re.escape(entry.entry_id)}-boiler-number-(?P<key>.+)$"
+    )
+    boiler_text_unique_id_re = re.compile(
+        rf"^{re.escape(entry.entry_id)}-boiler-text-(?P<key>phone_number)$"
+    )
+    circuit_sensor_unique_id_re = re.compile(
+        rf"^{re.escape(entry.entry_id)}-circuit-(?P<index>\d+)-sensor-(?P<key>.+)$"
+    )
+    circuit_number_unique_id_re = re.compile(
+        rf"^{re.escape(entry.entry_id)}-circuit-(?P<index>\d+)-number-(?P<key>.+)$"
+    )
+    adapter_hw_unique_id_re = re.compile(
+        rf"^{re.escape(entry.entry_id)}-adapter-hw-(?P<key>.+)$"
+    )
+    zone_demand_unique_id_re = re.compile(
+        rf"^{re.escape(entry.entry_id)}-zone-(?P<zone_id>.+)-heating-demand$"
+    )
+    zone_valve_unique_id_re = re.compile(
+        rf"^{re.escape(entry.entry_id)}-zone-(?P<zone_id>.+)-sensor-valve_position_pct$"
+    )
+    dhw_demand_unique_id = f"{entry.entry_id}-dhw-dhw-heating-demand"
+    system_text_unique_id_re = re.compile(
+        rf"^{re.escape(entry.entry_id)}-system-text-(?P<key>installer_name|installer_phone)$"
+    )
+
+    def _entry_entities() -> tuple:
+        return tuple(er.async_entries_for_config_entry(entity_registry, entry.entry_id))
+
+    def _device_has_entities(device_id: str) -> bool:
+        return any(entity_entry.device_id == device_id for entity_entry in entity_registry.entities.values())
+
+    def _current_fm5_payload() -> dict:
+        payload = fm5_coordinator.data if fm5_coordinator else None
+        return payload if isinstance(payload, dict) else {}
+
+    def _current_fm5_mode() -> str:
+        return str(_current_fm5_payload().get("fm5_semantic_mode") or "ABSENT").strip().upper()
+
+    def _current_radio_slots_with_live_values() -> dict[tuple[int, int], set[str]]:
+        payload = radio_coordinator.data if radio_coordinator else None
+        if not isinstance(payload, dict):
+            return {}
+        out: dict[tuple[int, int], set[str]] = {}
+        for radio in payload.get("radio_devices", []) or []:
+            if not isinstance(radio, dict):
+                continue
+            group = parse_optional_int(radio.get("group"))
+            instance = parse_optional_int(radio.get("instance"))
+            if group is None or instance is None:
+                continue
+            out[(group, instance)] = {key for key, value in radio.items() if value is not None}
+        return out
+
+    def _current_solar_live_keys() -> set[str]:
+        solar_payload = _current_fm5_payload().get("solar")
+        if not isinstance(solar_payload, dict):
+            return set()
+        return {key for key, value in solar_payload.items() if value is not None}
+
+    def _current_cylinder_live_keys() -> dict[int, set[str]]:
+        out: dict[int, set[str]] = {}
+        for cylinder in _current_fm5_payload().get("cylinders", []) or []:
+            if not isinstance(cylinder, dict):
+                continue
+            index = parse_optional_int(cylinder.get("index"))
+            if index is None or index < 0:
+                continue
+            out[index] = {key for key, value in cylinder.items() if value is not None}
+        return out
+
+    def _live_keys(payload: object) -> set[str]:
+        if not isinstance(payload, dict):
+            return set()
+        return {key for key, value in payload.items() if value is not None}
+
+    def _current_status_live_keys(target: str) -> set[str]:
+        payload = status_coordinator.data if status_coordinator else None
+        bucket = payload.get(target) if isinstance(payload, dict) else None
+        return _live_keys(bucket)
+
+    def _current_boiler_bucket_live_keys(bucket_name: str) -> set[str]:
+        payload = boiler_coordinator.data if boiler_coordinator else None
+        boiler_status = payload.get("boiler_status") if isinstance(payload, dict) else None
+        bucket = boiler_status.get(bucket_name) if isinstance(boiler_status, dict) else None
+        return _live_keys(bucket)
+
+    def _current_circuit_live_keys(bucket_name: str, index: int) -> set[str]:
+        payload = circuit_coordinator.data if circuit_coordinator else None
+        if not isinstance(payload, dict):
+            return set()
+        for circuit in payload.get("circuits", []) or []:
+            if not isinstance(circuit, dict):
+                continue
+            if parse_optional_int(circuit.get("index")) != index:
+                continue
+            return _live_keys(circuit.get(bucket_name))
+        return set()
+
+    def _current_adapter_hw_live_keys() -> set[str]:
+        payload = adapter_info_coordinator.data if adapter_info_coordinator else None
+        return _live_keys(payload)
+
+    def _current_zone_state_live_keys(zone_id: str) -> set[str]:
+        payload = semantic_coordinator.data if semantic_coordinator else None
+        if not isinstance(payload, dict):
+            return set()
+        for zone in payload.get("zones", []) or []:
+            if not isinstance(zone, dict):
+                continue
+            if str(zone.get("id")) != zone_id:
+                continue
+            return _live_keys(zone.get("state"))
+        return set()
+
+    def _current_dhw_state_live_keys() -> set[str]:
+        payload = semantic_coordinator.data if semantic_coordinator else None
+        dhw = payload.get("dhw") if isinstance(payload, dict) else None
+        state = dhw.get("state") if isinstance(dhw, dict) else None
+        return _live_keys(state)
+
+    def _current_system_config_live_keys() -> set[str]:
+        payload = system_coordinator.data if system_coordinator else None
+        config = payload.get("config") if isinstance(payload, dict) else None
+        return _live_keys(config)
+
+    def _is_sparse_entity_candidate(unique_id: str | None) -> bool:
+        if not unique_id:
+            return False
+        return (
+            unique_id in status_sparse_unique_ids
+            or unique_id == dhw_demand_unique_id
+            or radio_sensor_unique_id_re.match(unique_id) is not None
+            or solar_unique_id_re.match(unique_id) is not None
+            or cylinder_unique_id_re.match(unique_id) is not None
+            or boiler_temp_unique_id_re.match(unique_id) is not None
+            or boiler_sensor_sparse_unique_id_re.match(unique_id) is not None
+            or boiler_number_unique_id_re.match(unique_id) is not None
+            or boiler_text_unique_id_re.match(unique_id) is not None
+            or circuit_sensor_unique_id_re.match(unique_id) is not None
+            or circuit_number_unique_id_re.match(unique_id) is not None
+            or adapter_hw_unique_id_re.match(unique_id) is not None
+            or zone_demand_unique_id_re.match(unique_id) is not None
+            or zone_valve_unique_id_re.match(unique_id) is not None
+            or system_text_unique_id_re.match(unique_id) is not None
+        )
+
+    def _is_sparse_entity_live(unique_id: str | None) -> bool:
+        if not unique_id:
+            return False
+        if unique_id in status_sparse_unique_ids:
+            return "admission_repair_code" in _current_status_live_keys("daemon")
+        if unique_id == dhw_demand_unique_id:
+            return "heating_demand_pct" in _current_dhw_state_live_keys()
+
+        boiler_temp_match = boiler_temp_unique_id_re.match(unique_id)
+        if boiler_temp_match:
+            return boiler_temp_match.group("key") in _current_boiler_bucket_live_keys("state")
+
+        boiler_sensor_match = boiler_sensor_sparse_unique_id_re.match(unique_id)
+        if boiler_sensor_match:
+            return boiler_sensor_match.group("key") in _current_boiler_bucket_live_keys("config")
+
+        boiler_number_match = boiler_number_unique_id_re.match(unique_id)
+        if boiler_number_match:
+            return boiler_number_match.group("key") in _current_boiler_bucket_live_keys("config")
+
+        boiler_text_match = boiler_text_unique_id_re.match(unique_id)
+        if boiler_text_match:
+            return boiler_text_match.group("key") in _current_boiler_bucket_live_keys("config")
+
+        circuit_sensor_match = circuit_sensor_unique_id_re.match(unique_id)
+        if circuit_sensor_match:
+            index = int(circuit_sensor_match.group("index"))
+            return circuit_sensor_match.group("key") in _current_circuit_live_keys("state", index)
+
+        circuit_number_match = circuit_number_unique_id_re.match(unique_id)
+        if circuit_number_match:
+            index = int(circuit_number_match.group("index"))
+            return circuit_number_match.group("key") in _current_circuit_live_keys("config", index)
+
+        adapter_hw_match = adapter_hw_unique_id_re.match(unique_id)
+        if adapter_hw_match:
+            return adapter_hw_match.group("key") in _current_adapter_hw_live_keys()
+
+        zone_demand_match = zone_demand_unique_id_re.match(unique_id)
+        if zone_demand_match:
+            return "heating_demand_pct" in _current_zone_state_live_keys(zone_demand_match.group("zone_id"))
+
+        zone_valve_match = zone_valve_unique_id_re.match(unique_id)
+        if zone_valve_match:
+            return "valve_position_pct" in _current_zone_state_live_keys(zone_valve_match.group("zone_id"))
+
+        system_text_match = system_text_unique_id_re.match(unique_id)
+        if system_text_match:
+            return system_text_match.group("key") in _current_system_config_live_keys()
+
+        radio_match = radio_sensor_unique_id_re.match(unique_id)
+        if radio_match:
+            radio_slots_with_live_values = _current_radio_slots_with_live_values()
+            slot = (
+                int(radio_match.group("group"), 16),
+                int(radio_match.group("instance")),
+            )
+            return radio_match.group("key") in radio_slots_with_live_values.get(slot, set())
+
+        solar_match = solar_unique_id_re.match(unique_id)
+        if solar_match:
+            if _current_fm5_mode() != "INTERPRETED":
+                return False
+            solar_live_keys = _current_solar_live_keys()
+            return solar_match.group("key") in solar_live_keys
+
+        cylinder_match = cylinder_unique_id_re.match(unique_id)
+        if cylinder_match:
+            if _current_fm5_mode() != "INTERPRETED":
+                return False
+            cylinder_live_keys = _current_cylinder_live_keys()
+            index = int(cylinder_match.group("index"))
+            live_keys = cylinder_live_keys.get(index, set())
+            config_key = cylinder_match.group("config_key")
+            if config_key:
+                return config_key in live_keys
+            return "temperature_c" in live_keys
+        return False
+
+    def _radio_bus_key_from_unique_id(unique_id: str | None) -> str | None:
+        if not unique_id:
+            return None
+        for pattern in (radio_sensor_unique_id_re, radio_connected_unique_id_re):
+            match = pattern.match(unique_id)
+            if match is None:
+                continue
+            return build_radio_bus_key(
+                int(match.group("group"), 16),
+                int(match.group("instance")),
+            )
+        return None
+
+    # ADR-027: build set of merged B524 slot prefixes for entity cleanup.
+    # bus_key format: "g0c-i01" -> unique_id slot format: "0c-01"
+    def _bus_key_to_uid_slot(bk: str) -> str:
+        return bk.replace("g", "").replace("-i", "-")
+
+    _b524_merged_uid_prefixes = frozenset(
+        f"{entry.entry_id}-radio-{_bus_key_to_uid_slot(bus_key)}-sensor-"
+        for bus_key in b524_merge_targets
+    )
+
+    def cleanup_obsolete_entity_registry_entries() -> None:
+        removed = 0
+        for entity_entry in _entry_entities():
+            if entity_entry.platform != DOMAIN:
+                continue
+            unique_id = str(entity_entry.unique_id or "")
+            remove_entry = False
+            if entity_entry.domain in {"fan", "valve", "switch"}:
+                remove_entry = True
+            elif unique_id in status_sparse_unique_ids and not _is_sparse_entity_live(unique_id):
+                remove_entry = True
+            elif _stale_bus_address_unique_id(unique_id, entry.entry_id, known_bus_devices):
+                remove_entry = True
+            elif entity_entry.domain == "number" and re.match(
+                rf"^{re.escape(entry.entry_id)}-cylinder-\d+-number-",
+                unique_id,
+            ):
+                remove_entry = True
+            # ADR-027: remove redundant sensor entities from merged B524 slots.
+            elif entity_entry.domain == "sensor" and any(
+                unique_id.startswith(prefix) for prefix in _b524_merged_uid_prefixes
+            ):
+                remove_entry = True
+            elif (radio_bus_key := _radio_bus_key_from_unique_id(unique_id)) is not None:
+                remove_entry = should_remove_missing_radio_bus_key(
+                    radio_bus_key,
+                    known_radio_bus_keys,
+                    set(b524_merge_targets),
+                )
+            else:
+                cylinder_match = cylinder_unique_id_re.match(unique_id)
+                if cylinder_match and int(cylinder_match.group("index")) not in known_cylinder_indexes:
+                    remove_entry = True
+            if remove_entry:
+                entity_registry.async_remove(entity_entry.entity_id)
+                removed += 1
+        if removed:
+            _LOGGER.info(
+                "Helianthus cleanup removed %d obsolete entities for entry %s",
+                removed,
+                entry.entry_id,
+            )
+
+    def cleanup_nonexportable_radio_registry_entries(
+        radio_keys: set[str],
+        reason: str,
+    ) -> None:
+        if not radio_keys:
+            radio_keys = set()
+        radio_keys_to_remove = set(radio_keys)
+        removed_entities = 0
+        for entity_entry in _entry_entities():
+            if entity_entry.platform != DOMAIN:
+                continue
+            unique_id = str(entity_entry.unique_id or "")
+            radio_bus_key = _radio_bus_key_from_unique_id(unique_id)
+            if radio_bus_key is None:
+                continue
+            if radio_bus_key not in radio_keys_to_remove and not should_remove_missing_radio_bus_key(
+                radio_bus_key,
+                known_radio_bus_keys,
+                set(b524_merge_targets),
+            ):
+                continue
+            entity_registry.async_remove(entity_entry.entity_id)
+            removed_entities += 1
+
+        removed_devices = 0
+        radio_prefix = f"{entry.entry_id}-radio-"
+        for device_entry in tuple(dr.async_entries_for_config_entry(device_registry, entry.entry_id)):
+            remove_device = False
+            for identifier_domain, token in _iter_identifier_pairs(device_entry.identifiers):
+                if identifier_domain != DOMAIN or not token.startswith(radio_prefix):
+                    continue
+                radio_bus_key = token[len(radio_prefix):]
+                if radio_bus_key in radio_keys_to_remove or should_remove_missing_radio_bus_key(
+                    radio_bus_key,
+                    known_radio_bus_keys,
+                    set(b524_merge_targets),
+                ):
+                    remove_device = True
+                    break
+            if not remove_device:
+                continue
+            if _device_has_entities(device_entry.id):
+                continue
+            device_registry.async_remove_device(device_entry.id)
+            removed_devices += 1
+
+        if removed_entities or removed_devices:
+            _LOGGER.info(
+                "Helianthus cleanup removed %d entities and %d devices for non-exportable radio slots %s (%s)",
+                removed_entities,
+                removed_devices,
+                sorted(radio_keys),
+                reason,
+            )
+
+    def cleanup_obsolete_devices(reason: str) -> None:
+        # Remove stale calendar entities with old zone_N unique_id format
+        # (pre-fix: B555 zone indices were used directly instead of semantic zone IDs).
+        _stale_schedule_re = re.compile(
+            rf"^{re.escape(entry.entry_id)}-schedule-zone_\d+-"
+        )
+        stale_removed = 0
+        for entity_entry in tuple(_entry_entities()):
+            uid = str(entity_entry.unique_id or "")
+            if _stale_schedule_re.match(uid):
+                entity_registry.async_remove(entity_entry.entity_id)
+                stale_removed += 1
+        if stale_removed:
+            _LOGGER.info(
+                "Helianthus cleanup removed %d stale schedule entities for entry %s (%s)",
+                stale_removed,
+                entry.entry_id,
+                reason,
+            )
+
+        removed = 0
+        for device_entry in tuple(dr.async_entries_for_config_entry(device_registry, entry.entry_id)):
+            identifier_pairs = _iter_identifier_pairs(device_entry.identifiers)
+            domain_tokens = [
+                token for identifier_domain, token in identifier_pairs if identifier_domain == DOMAIN
+            ]
+            if not domain_tokens:
+                continue
+            remove_device = False
+            for token in domain_tokens:
+                if token in {
+                    f"{entry.entry_id}-boiler-burner",
+                    f"{entry.entry_id}-boiler-hydraulics",
+                }:
+                    remove_device = True
+                    break
+                if _is_stale_bus_identifier(token, entry.entry_id, known_bus_devices):
+                    remove_device = True
+                    break
+                if token.startswith(f"{entry.entry_id}-zone-"):
+                    remove_device = True
+                    break
+                cylinder_index = _parse_identifier_index(token, f"{entry.entry_id}-cylinder-")
+                if cylinder_index is not None and cylinder_index not in known_cylinder_indexes:
+                    remove_device = True
+                    break
+                # ADR-027: remove shadow radio device for merged B524 function modules.
+                radio_prefix = f"{entry.entry_id}-radio-"
+                if token.startswith(radio_prefix):
+                    radio_bus_key = token[len(radio_prefix):]
+                    if should_remove_missing_radio_bus_key(
+                        radio_bus_key,
+                        known_radio_bus_keys,
+                        set(b524_merge_targets),
+                    ):
+                        remove_device = True
+                        break
+            if not remove_device:
+                continue
+            if _device_has_entities(device_entry.id):
+                continue
+            device_registry.async_remove_device(device_entry.id)
+            removed += 1
+        if removed:
+            _LOGGER.info(
+                "Helianthus cleanup removed %d obsolete devices for entry %s (%s)",
+                removed,
+                entry.entry_id,
+                reason,
+            )
+
+    def auto_enable_sparse_entities(reason: str) -> None:
+        registry_disabler = getattr(er, "RegistryEntryDisabler", None)
+        integration_disabler = "integration"
+        if registry_disabler is not None:
+            integration_disabler = getattr(
+                registry_disabler,
+                "INTEGRATION",
+                integration_disabler,
+            )
+        existing_unique_ids = {
+            str(entity_entry.unique_id or "")
+            for entity_entry in _entry_entities()
+            if entity_entry.platform == DOMAIN
+        }
+        missing_live_status_ids = [
+            unique_id
+            for unique_id in status_sparse_unique_ids
+            if unique_id not in existing_unique_ids and _is_sparse_entity_live(unique_id)
+        ]
+        if missing_live_status_ids:
+            _LOGGER.info(
+                "Helianthus scheduling reload for %d live sparse status entities for entry %s (%s)",
+                len(missing_live_status_ids),
+                entry.entry_id,
+                reason,
+            )
+            schedule_reload("sparse status entities became live")
+            return
+        enabled = 0
+        for entity_entry in _entry_entities():
+            if entity_entry.platform != DOMAIN:
+                continue
+            if entity_entry.disabled_by != integration_disabler:
+                continue
+            if not _is_sparse_entity_live(str(entity_entry.unique_id or "")):
+                continue
+            entity_registry.async_update_entity(entity_entry.entity_id, disabled_by=None)
+            enabled += 1
+        if enabled:
+            _LOGGER.info(
+                "Helianthus auto-enabled %d sparse entities for entry %s (%s)",
+                enabled,
+                entry.entry_id,
+                reason,
+            )
+            schedule_reload("sparse entities became live")
+
+    def auto_disable_sparse_entities(reason: str) -> None:
+        registry_disabler = getattr(er, "RegistryEntryDisabler", None)
+        integration_disabler = "integration"
+        if registry_disabler is not None:
+            integration_disabler = getattr(
+                registry_disabler,
+                "INTEGRATION",
+                integration_disabler,
+            )
+        disabled = 0
+        for entity_entry in _entry_entities():
+            if entity_entry.platform != DOMAIN:
+                continue
+            if entity_entry.disabled_by is not None:
+                continue
+            unique_id = str(entity_entry.unique_id or "")
+            if not _is_sparse_entity_candidate(unique_id):
+                continue
+            if _is_sparse_entity_live(unique_id):
+                continue
+            entity_registry.async_update_entity(
+                entity_entry.entity_id,
+                disabled_by=integration_disabler,
+            )
+            disabled += 1
+        if disabled:
+            _LOGGER.info(
+                "Helianthus auto-disabled %d sparse entities for entry %s (%s)",
+                disabled,
+                entry.entry_id,
+                reason,
+            )
+            schedule_reload("sparse entities became null")
+
+    def run_trusted_cleanup(reason: str) -> None:
+        nonlocal cleanup_ran
+        if cleanup_ran:
+            return
+        if not cleanup_allowed:
+            _LOGGER.info(
+                "Helianthus cleanup skipped for entry %s until admission and inventory are trusted (%s)",
+                entry.entry_id,
+                admission.get("repair_code") or REPAIR_EMPTY_INVENTORY_UNTRUSTED,
+            )
+            return
+        cleanup_obsolete_entity_registry_entries()
+        auto_disable_sparse_entities(reason)
+        auto_enable_sparse_entities(reason)
+        cleanup_obsolete_devices(reason)
+        cleanup_ran = True
+
+    def existing_zone_parent_device_ids(zones: list[dict]) -> dict[str, tuple[str, str]]:
+        """Recover validated radio parents from existing climate registry entries."""
+        zone_ids_by_unique_id = zone_ids_by_climate_unique_id(entry.entry_id, zones)
+        radio_identifier_prefix = f"{entry.entry_id}-radio-"
+        existing: dict[str, tuple[str, str]] = {}
+        for entity_entry in _entry_entities():
+            zone_id = zone_ids_by_unique_id.get(entity_entry.unique_id)
+            if zone_id is None or entity_entry.device_id is None:
+                continue
+            device_entry = device_registry.async_get(entity_entry.device_id)
+            if device_entry is None:
+                continue
+            for identifier in sorted(device_entry.identifiers):
+                if identifier[0] == DOMAIN and identifier[1].startswith(radio_identifier_prefix):
+                    existing[zone_id] = identifier
+                    break
+        return existing
+
+    initial_semantic_payload = (
+        semantic_coordinator.data if isinstance(semantic_coordinator.data, dict) else {}
+    )
+    initial_zones = [
+        zone
+        for zone in (initial_semantic_payload.get("zones", []) or [])
+        if isinstance(zone, dict)
+    ]
+    initial_radio_payload = (
+        radio_coordinator.data if isinstance(radio_coordinator.data, dict) else None
+    )
+    initial_live_parent_ids, _ = build_zone_parent_device_ids(
+        entry.entry_id,
+        initial_zones,
+        initial_radio_payload,
+        regulator_device,
+    )
+    initial_mappings = radio_mappings_by_zone_id(initial_zones)
+    binding_key = "zone_parent_bindings_v1"
+    retained_zone_parent_ids, retained_zone_parent_mappings, normalized_bindings = (
+        resolve_retained_parent_bindings(
+            existing_zone_parent_device_ids(initial_zones),
+            entry.data.get(binding_key),
+        )
+    )
+    for zone_id, parent_id in initial_live_parent_ids.items():
+        mapping = initial_mappings.get(zone_id)
+        if mapping is None:
+            continue
+        retained_zone_parent_ids[zone_id] = parent_id
+        retained_zone_parent_mappings[zone_id] = mapping
+        normalized_bindings[zone_id] = {
+            "identifier": parent_id[1],
+            "mapping": mapping,
+        }
+
+    def persist_retained_zone_parent_bindings() -> None:
+        if entry.data.get(binding_key) == normalized_bindings:
+            return
+        updated_data = dict(entry.data)
+        updated_data[binding_key] = dict(normalized_bindings)
+        hass.config_entries.async_update_entry(entry, data=updated_data)
+
+    persist_retained_zone_parent_bindings()
+
+    def adopt_current_live_parent_bindings() -> None:
+        payload = semantic_coordinator.data if isinstance(semantic_coordinator.data, dict) else {}
+        zones = [zone for zone in (payload.get("zones", []) or []) if isinstance(zone, dict)]
+        radio_payload = radio_coordinator.data if isinstance(radio_coordinator.data, dict) else None
+        live_parent_ids, _ = build_zone_parent_device_ids(
+            entry.entry_id,
+            zones,
+            radio_payload,
+            regulator_device,
+        )
+        mappings = radio_mappings_by_zone_id(zones)
+        changed = False
+        for zone_id, parent_id in live_parent_ids.items():
+            mapping = mappings.get(zone_id)
+            if mapping is None:
+                continue
+            binding = {"identifier": parent_id[1], "mapping": mapping}
+            if normalized_bindings.get(zone_id) == binding:
+                continue
+            retained_zone_parent_ids[zone_id] = parent_id
+            retained_zone_parent_mappings[zone_id] = mapping
+            normalized_bindings[zone_id] = binding
+            changed = True
+        if changed:
+            persist_retained_zone_parent_bindings()
+
+    def current_zone_parent_device_ids() -> tuple[dict[str, tuple[str, str]], tuple[str, ...]]:
+        payload = semantic_coordinator.data if isinstance(semantic_coordinator.data, dict) else {}
+        zones = [zone for zone in (payload.get("zones", []) or []) if isinstance(zone, dict)]
+        radio_payload = radio_coordinator.data if isinstance(radio_coordinator.data, dict) else None
+        return build_zone_parent_device_ids(
+            entry.entry_id,
+            zones,
+            radio_payload,
+            regulator_device,
+            existing_parent_device_ids=retained_zone_parent_ids,
+            existing_parent_mappings=retained_zone_parent_mappings,
+            allow_existing_parent_fallback=(
+                isinstance(radio_payload, dict)
+                and radio_payload.get("inventory_complete") is False
+            ),
+        )
+
+    zone_parent_device_ids, unresolved_zone_ids = current_zone_parent_device_ids()
+    if unresolved_zone_ids:
+        _LOGGER.warning(
+            "Helianthus zone parent resolution incomplete for entry %s; skipping unresolved zones until reload: %s",
+            entry.entry_id,
+            ", ".join(unresolved_zone_ids),
+        )
+
+    # Reverse map: device_id → zone_name for radio device naming (HA 2026.4).
+    radio_device_zone_names: dict[tuple[str, str], str] = {}
+    if isinstance(semantic_coordinator.data, dict):
+        for zone in semantic_coordinator.data.get("zones", []) or []:
+            if not isinstance(zone, dict):
+                continue
+            zone_id = zone.get("id")
+            zone_name = zone.get("name")
+            if zone_id is None or not zone_name:
+                continue
+            from .climate import _normalize_zone_id as _nzid
+            normalized = _nzid(zone_id)
+            if normalized is None:
+                continue
+            device_id = zone_parent_device_ids.get(normalized)
+            if device_id is not None and device_id != regulator_device:
+                radio_device_zone_names[device_id] = str(zone_name).strip()
+
+    zone_schedule_helpers = _parse_zone_schedule_helper_bindings(
+        str(
+            entry.options.get(
+                CONF_ZONE_SCHEDULE_HELPERS,
+                DEFAULT_ZONE_SCHEDULE_HELPERS,
+            )
+        )
+    )
+    dhw_schedule_helper = str(
+        entry.options.get(CONF_DHW_SCHEDULE_HELPER, DEFAULT_DHW_SCHEDULE_HELPER) or ""
+    ).strip()
+
+    async def apply_zone_schedule(zone_id: str) -> None:
+        if not status_admission_trusted(status_coordinator):
+            _LOGGER.warning("Zone schedule helper ignored: source admission is not trusted")
+            return
+        if regulator_bus_address is None:
+            _LOGGER.warning("Zone schedule helper ignored: regulator address missing")
+            return
+        instance = _zone_instance_from_id(zone_id)
+        if instance is None:
+            _LOGGER.warning("Zone schedule helper ignored: invalid zone id %s", zone_id)
+            return
+        variables = {
+            "address": int(regulator_bus_address),
+            "params": {
+                "opcode": 0x02,
+                "group": 0x03,
+                "instance": int(instance),
+                "addr": 0x0006,
+                "data": [1, 0x00],
+            },
+        }
+        try:
+            payload = await client.mutation(_INVOKE_SET_EXT_REGISTER, variables)
+            invoke = payload.get("invoke") if isinstance(payload, dict) else None
+            if not isinstance(invoke, dict) or not invoke.get("ok"):
+                _LOGGER.warning("Zone schedule helper write failed for %s: %s", zone_id, invoke)
+                return
+            await semantic_coordinator.async_request_refresh()
+        except Exception as exc:  # pragma: no cover - defensive runtime guard
+            _LOGGER.warning("Zone schedule helper write failed for %s: %s", zone_id, exc)
+
+    async def apply_dhw_schedule() -> None:
+        if not status_admission_trusted(status_coordinator):
+            _LOGGER.warning("DHW schedule helper ignored: source admission is not trusted")
+            return
+        if regulator_bus_address is None:
+            _LOGGER.warning("DHW schedule helper ignored: regulator address missing")
+            return
+        variables = {
+            "address": int(regulator_bus_address),
+            "params": {
+                "opcode": 0x02,
+                "group": 0x01,
+                "instance": 0x00,
+                "addr": 0x0003,
+                "data": [1, 0x00],
+            },
+        }
+        try:
+            payload = await client.mutation(_INVOKE_SET_EXT_REGISTER, variables)
+            invoke = payload.get("invoke") if isinstance(payload, dict) else None
+            if not isinstance(invoke, dict) or not invoke.get("ok"):
+                _LOGGER.warning("DHW schedule helper write failed: %s", invoke)
+                return
+            await semantic_coordinator.async_request_refresh()
+        except Exception as exc:  # pragma: no cover - defensive runtime guard
+            _LOGGER.warning("DHW schedule helper write failed: %s", exc)
+
+    def current_bus_device_ids(current: list[dict]) -> set[str]:
+        current_ids: set[str] = set()
+        for dev in current:
+            bus_key = resolved_bus_device_key(dev)
+            if bus_key:
+                current_ids.add(bus_key)
+        return current_ids
+
+    def handle_device_update() -> None:
+        current = device_coordinator.data or []
+        record_cleanup_evidence(current)
+        if not current:
+            return
+        current_ids = current_bus_device_ids(current)
+        if not current_ids:
+            return
+        if current_ids.issubset(known_bus_devices):
+            if cleanup_allowed:
+                run_trusted_cleanup("trusted non-empty inventory refresh")
+            return
+        schedule_reload("new bus devices discovered")
+
+    def handle_status_update() -> None:
+        current = device_coordinator.data or []
+        record_cleanup_evidence(current)
+        current_ids = current_bus_device_ids(current)
+        if current_ids and not current_ids.issubset(known_bus_devices):
+            schedule_reload("new bus devices discovered")
+            return
+        auto_enable_sparse_entities("status update")
+        if cleanup_allowed and current_ids:
+            run_trusted_cleanup("trusted admission status refresh")
+
+    def handle_semantic_update() -> None:
+        payload = semantic_coordinator.data or {}
+        zones = payload.get("zones", []) or []
+        dhw = payload.get("dhw")
+        current_zone_ids: set[str] = {
+            str(zone.get("id")) for zone in zones if zone.get("id") is not None
+        }
+        has_new_zones = bool(current_zone_ids - known_zones)
+        has_new_dhw = dhw is not None and not known_has_dhw
+        adopt_current_live_parent_bindings()
+        current_zone_parent_ids, unresolved = current_zone_parent_device_ids()
+        if should_reload_zone_parent_state(
+            zone_parent_device_ids,
+            unresolved_zone_ids,
+            current_zone_parent_ids,
+            unresolved,
+        ):
+            schedule_reload("zone parent mapping changed")
+            return
+        if has_new_zones or has_new_dhw:
+            schedule_reload("semantic inventory became available")
+            return
+        auto_enable_sparse_entities("semantic update")
+
+    def handle_circuit_update() -> None:
+        payload = circuit_coordinator.data or {}
+        current_indexes: set[int] = set()
+        for circuit in payload.get("circuits", []) or []:
+            if not isinstance(circuit, dict):
+                continue
+            index = parse_circuit_index(circuit)
+            if index is not None:
+                current_indexes.add(index)
+        if current_indexes - known_circuit_indexes:
+            schedule_reload("circuit inventory became available")
+            return
+        auto_enable_sparse_entities("circuit update")
+
+    def handle_radio_update() -> None:
+        nonlocal nonexportable_radio_keys
+        payload = radio_coordinator.data or {}
+        radio_devices = payload.get("radio_devices", []) or []
+        current_nonexportable = nonexportable_radio_bus_keys(radio_devices)
+        nonexportable_radio_keys = current_nonexportable
+        if current_nonexportable:
+            cleanup_nonexportable_radio_registry_entries(
+                current_nonexportable,
+                "radio update",
+            )
+        current_keys = exportable_radio_bus_keys(radio_devices)
+        adopt_current_live_parent_bindings()
+        current_zone_parent_ids, unresolved = current_zone_parent_device_ids()
+        if should_reload_zone_parent_state(
+            zone_parent_device_ids,
+            unresolved_zone_ids,
+            current_zone_parent_ids,
+            unresolved,
+        ):
+            schedule_reload("zone parent mapping changed")
+            return
+        auto_enable_sparse_entities("radio update")
+        if current_keys - known_radio_bus_keys:
+            schedule_reload("radio inventory became available")
+
+    def handle_fm5_update() -> None:
+        payload = fm5_coordinator.data or {}
+        mode = str(payload.get("fm5_semantic_mode") or "ABSENT").strip().upper()
+        current_indexes: set[int] = set()
+        for cylinder in payload.get("cylinders", []) or []:
+            if not isinstance(cylinder, dict):
+                continue
+            index = parse_optional_int(cylinder.get("index"))
+            if index is not None and index >= 0:
+                current_indexes.add(index)
+        if mode != known_fm5_mode:
+            if "INTERPRETED" in {mode, known_fm5_mode}:
+                schedule_reload("fm5 semantic mode changed")
+                return
+        if mode == "INTERPRETED" and (current_indexes - known_cylinder_indexes):
+            schedule_reload("cylinder inventory became available")
+            return
+        auto_enable_sparse_entities("fm5 update")
+
+    def handle_boiler_update() -> None:
+        auto_enable_sparse_entities("boiler update")
+
+    def handle_adapter_info_update() -> None:
+        auto_enable_sparse_entities("adapter info update")
+
+    unsub_listeners: list[Callable[[], None]] = []
+    unsub_listeners.append(status_coordinator.async_add_listener(handle_status_update))
+    unsub_listeners.append(device_coordinator.async_add_listener(handle_device_update))
+    unsub_listeners.append(semantic_coordinator.async_add_listener(handle_semantic_update))
+    unsub_listeners.append(circuit_coordinator.async_add_listener(handle_circuit_update))
+    unsub_listeners.append(radio_coordinator.async_add_listener(handle_radio_update))
+    unsub_listeners.append(fm5_coordinator.async_add_listener(handle_fm5_update))
+    unsub_listeners.append(boiler_coordinator.async_add_listener(handle_boiler_update))
+    unsub_listeners.append(adapter_info_coordinator.async_add_listener(handle_adapter_info_update))
+
+    for zone_id, helper_entity in zone_schedule_helpers.items():
+        @callback
+        def _handle_zone_schedule_event(event, zone_id=zone_id, helper_entity=helper_entity):
+            if str(event.data.get("entity_id") or "") != helper_entity:
+                return
+            new_state = event.data.get("new_state")
+            if new_state is None:
+                return
+            if str(new_state.state).strip().lower() != "on":
+                return
+            hass.async_create_task(apply_zone_schedule(zone_id))
+
+        unsub_listeners.append(async_track_state_change_event(hass, [helper_entity], _handle_zone_schedule_event))
+
+    if dhw_schedule_helper.startswith("schedule."):
+        @callback
+        def _handle_dhw_schedule_event(event):
+            new_state = event.data.get("new_state")
+            if new_state is None:
+                return
+            if str(new_state.state).strip().lower() != "on":
+                return
+            hass.async_create_task(apply_dhw_schedule())
+
+        unsub_listeners.append(
+            async_track_state_change_event(hass, [dhw_schedule_helper], _handle_dhw_schedule_event)
+        )
+
+    use_subscriptions = entry.options.get(CONF_USE_SUBSCRIPTIONS, DEFAULT_USE_SUBSCRIPTIONS)
+    subscription_task = None
+    if use_subscriptions:
+        boiler_subscription_coordinator = (
+            boiler_coordinator if getattr(boiler_coordinator, "boiler_supported", False) else None
+        )
+        subscription_task = await start_subscriptions(
+            session,
+            graphql_url,
+            semantic_coordinator,
+            energy_coordinator,
+            boiler_subscription_coordinator,
+            radio_coordinator,
+        )
+
+    boiler_physical_device_id = resolve_boiler_physical_device_id(
+        boiler_device,
+        regulator_device,
+    )
+    boiler_via_device_id = resolve_boiler_via_device_id(
+        boiler_device,
+        regulator_device,
+        adapter_device_id,
+    )
+    # HA-1 reduced profile: boiler Burner/Hydraulics sub-devices are not registered yet.
+    boiler_burner_device_id = boiler_burner_identifier(entry.entry_id)
+    boiler_hydraulics_device_id = boiler_hydraulics_identifier(entry.entry_id)
+
+    pv_m2m_boundary = None
+    pv_m2m_coordinator = None
+    try:
+        pv_m2m_boundary = await async_setup_pv_m2m_boundary(
+            hass,
+            entry,
+            scan_interval=scan_interval,
+        )
+    except (OSError, TypeError, ValueError):
+        _LOGGER.warning(
+            "Canonical PV consumer configuration failed for entry %s",
+            entry.entry_id,
+        )
+    if pv_m2m_boundary is not None:
+        pv_m2m_coordinator = pv_m2m_boundary.coordinator
+
+    hass.data.setdefault(DOMAIN, {})[entry.entry_id] = {
+        "device_coordinator": device_coordinator,
+        "status_coordinator": status_coordinator,
+        "semantic_coordinator": semantic_coordinator,
+        "energy_coordinator": energy_coordinator,
+        "circuit_coordinator": circuit_coordinator,
+        "radio_coordinator": radio_coordinator,
+        "fm5_coordinator": fm5_coordinator,
+        "system_coordinator": system_coordinator,
+        "boiler_coordinator": boiler_coordinator,
+        "schedule_coordinator": schedule_coordinator,
+        "adapter_info_coordinator": adapter_info_coordinator,
+        "eebus_admin_coordinator": admin_coordinator,
+        "eebus_admin_action_broker": admin_coordinator.lifecycle.action_broker,
+        "eebus_admin_session": admin_session,
+        "eebus_admin_services_registered": admin_services_registered,
+        "pv_m2m_coordinator": pv_m2m_coordinator,
+        "pv_m2m_boundary": pv_m2m_boundary,
+        "pv_m2m_option_signature": pv_m2m_option_signature(entry.options),
+        "graphql_client": client,
+        "subscription_task": subscription_task,
+        "unsub_listeners": unsub_listeners,
+        "daemon_device_id": daemon_device_id,
+        "adapter_device_id": adapter_device_id,
+        "boiler_device_id": boiler_device,
+        "regulator_device_id": regulator_device,
+        "vr71_device_id": vr71_device,
+        "regulator_manufacturer": regulator_manufacturer,
+        "regulator_bus_address": regulator_bus_address,
+        "admission_repair_code": admission.get("repair_code"),
+        "source_selection": admission.get("source_selection"),
+        "boiler_physical_device_id": boiler_physical_device_id,
+        "boiler_via_device_id": boiler_via_device_id,
+        "boiler_burner_device_id": boiler_burner_device_id,
+        "boiler_hydraulics_device_id": boiler_hydraulics_device_id,
+        "zone_parent_device_ids": zone_parent_device_ids,
+        "radio_device_zone_names": radio_device_zone_names,
+        "unresolved_zone_ids": unresolved_zone_ids,
+        "b524_merge_targets": b524_merge_targets,
+    }
+
+    async def _reload_on_pv_m2m_option_change(
+        updated_hass: HomeAssistant,
+        updated_entry: ConfigEntry,
+    ) -> None:
+        runtime = updated_hass.data.get(DOMAIN, {}).get(updated_entry.entry_id)
+        if not isinstance(runtime, dict):
+            return
+        if runtime.get("pv_m2m_option_signature") == pv_m2m_option_signature(
+            updated_entry.options
+        ):
+            return
+        await updated_hass.config_entries.async_reload(updated_entry.entry_id)
+
+    if hasattr(entry, "add_update_listener") and hasattr(entry, "async_on_unload"):
+        entry.async_on_unload(entry.add_update_listener(_reload_on_pv_m2m_option_change))
+
+    # Run stale cleanup again after platform setup so registry-backed entities
+    # created by prior installs are removed from the fully materialized registry.
+    await _async_forward_platforms_and_finalize(
+        hass,
+        entry,
+        PLATFORMS,
+        cleanup_nonexportable_radio_registry_entries=cleanup_nonexportable_radio_registry_entries,
+        nonexportable_radio_keys=nonexportable_radio_keys,
+        run_trusted_cleanup=run_trusted_cleanup,
+    )
+
+    return True
